Add files via upload
 Use a Collection to store the objects of classes derived from the same super class.
 Use foreach to process the elements of a Collection.
 Validate the data what you get from the user; throw Exception for invalid data, and handle
the thrown Exceptions.

Choose a point on the plane, and fill a collection with several regular shapes (circle, regular
triangle, square, regular hexagon). Which shapes lies the closest to the point?
In case of the circle, the distance is measured from the line of the circle, if the point lies outside
of the circle. If a point lies inside of a circle, their distance is considered as zero.
Each shape can be represented by its center and side length (or radius), if we assume that one
side of the polygons are parallel with x axis, and its nodes lies on or above this side.
Load and create the shapes from a text file. The first line of the file contains the number of the
shapes, and each following line contain a shape. The first character will identify the type of the
shape, which is followed by the center coordinate and the side length or radius.
Manage the shapes uniformly, so derive them from the same super class.
 The documentation should contain:
o the description of the exercise,
o the class diagram,
o the short description of each methods,
o and the testing (white box / black box).